+++ descent.core/src/descent/internal/compiler/env/IGenericType.java
@@ -18,7 +18,7 @@ public interface IGenericType extends IDependent {
  * NOTE 1: We have added AccDeprecated & AccSynthetic.
  * NOTE 2: If the receiver represents a member type, the modifiers are extracted from its inner class attributes.
  */
-int getModifiers();
+	long getModifiers();
 /**
  * Answer whether the receiver contains the resolved binary form
  * or the unresolved source form of the type.
